add changes.

+++ app/model/article_model.go
@@ -3,10 +3,10 @@ package model
 type ArticleShareRequest struct {
 	BaseRequest
 	OpenId    string `json:"openId"`
-	ArticleId int    `json:"articleId"`
+	ArticleId int64    `json:"articleId"`
 }
 
 // Return Data field of article share response
 type ArticleShareReturnData struct {
-	ShareTimes int `json:"shareTimes"`
+	ShareTimes int32 `json:"shareTimes"`
 }
